refactor : 폴더 위치 및 파일 분리
