+++ README.md
@@ -29,8 +29,8 @@
 
 ## step2
 
-- [ ] 음식점 목록 카테고리 필터링
-  - [ ] 이름순/거리순으로 정렬 가능
+- [x] 음식점 목록 카테고리 필터링
+  - [x] 이름순/거리순으로 정렬 가능
 - [ ] 음식점 상새 정보
   - [ ] 카테고리, 이름, 거리, 설명, 참고 링크 제공
 - [ ] 음식점 삭제
