+++ src/components/textArea.js
@@ -1,8 +1,8 @@
 import { title } from "../constants/inputTitle.js";
+import createElement from "../utils/createElement.js";
 
 const TextArea = (name, helpText, colRow = { col: 30, row: 5 }) => {
-  const formItem = document.createElement("div");
-  formItem.classList.add("form-item");
+  const formItem = createElement({ type: "div", classList: ["form-item"] });
 
   formItem.innerHTML = `
   <label for="${name}">${title[name]}</label>
