--- src/utils/utilsUI.js
@@ -1,174 +1,0 @@
-import Button from "../components/Button.js";
-import Store from "../components/Store.js";
-import querySelector from "./querySelector.js";
-import validate from "./validate.js";
-import optionInput from "../components/OptionInput.js";
-import options from "../constants/options.js";
-import textInput from "../components/TextInput.js";
-import textArea from "../components/TextArea.js";
-import helpText from "../constants/helpText.js";
-import Header from "../components/Header.js";
-
-export const headerUtils = {
-  addHeader: (title) => {
-    const header = querySelector(".header");
-    header.innerHTML = Header(title);
-  },
-};
-
-export const modalUtils = {
-  closeModal: () => {
-    const modal = querySelector(".modal");
-    modal.classList.remove("modal--open");
-  },
-
-  addForm: () => {
-    const modalContainer = querySelector(".modal-container");
-    modalContainer.innerHTML = `<h2 class="modal-title text-title">새로운 음식점</h2>
-    <form class="modal-form"></form>`;
-
-    const modalForm = querySelector(".modal-form");
-    modalForm.appendChild(optionInput("category", options.category));
-    modalForm.appendChild(textInput("name", true));
-    modalForm.appendChild(optionInput("distance", options.distance));
-    modalForm.appendChild(textArea("description", helpText.description));
-    modalForm.appendChild(textInput("link", false, helpText.link));
-
-    modalForm.appendChild(modalUtils.addButtons());
-    modalUtils.addFormCheck();
-
-    querySelector("#cancel-button").addEventListener(
-      "click",
-      modalUtils.closeModal
-    );
-  },
-
-  addButtons: () => {
-    const buttonContainer = document.createElement("div");
-    buttonContainer.classList.add("button-container");
-    buttonContainer.appendChild(
-      Button({
-        name: "취소하기",
-        type: "button",
-        class: "button--secondary",
-        id: "cancel-button",
-      })
-    );
-    buttonContainer.appendChild(
-      Button({
-        name: "추가하기",
-        type: "submit",
-        class: "button--primary",
-        id: "add-button",
-      })
-    );
-
-    return buttonContainer;
-  },
-
-  addFormCheck: () => {
-    const nameInput = querySelector("#name");
-    const descInput = querySelector("#description");
-    const linkInput = querySelector("#link");
-    const categorySelect = querySelector("#category");
-    const distSelect = querySelector("#distance");
-
-    modalUtils.checkInput(nameInput, validate.nameLength);
-    modalUtils.checkInput(descInput, validate.descLength);
-    modalUtils.checkInput(linkInput, validate.linkForm);
-    modalUtils.checkInput(categorySelect, validate.emptySelector, "change");
-    modalUtils.checkInput(distSelect, validate.emptySelector, "change");
-  },
-
-  checkInput: (input, validate, type = "input") => {
-    const addButton = querySelector("#add-button");
-
-    input.addEventListener(type, (e) => {
-      try {
-        validate(e.target.value);
-        modalUtils.removeErrorText(input);
-        addButton.classList.remove("disabled-button");
-        addButton.disabled = false;
-      } catch (e) {
-        modalUtils.addErrorText(input, e);
-        addButton.classList.add("disabled-button");
-        addButton.disabled = true;
-      }
-    });
-  },
-
-  addErrorText: (input, e) => {
-    if (!input.classList.contains("form-item--error")) {
-      input.classList.add("form-item--error");
-      const parentNode = input.parentNode;
-      const errorText = document.createElement("span");
-      errorText.classList.add("error-text");
-      errorText.innerText = e.message;
-      parentNode.appendChild(errorText);
-    }
-  },
-
-  removeErrorText: (input) => {
-    if (input.parentNode.querySelector(".error-text")) {
-      input.classList.remove("form-item--error");
-      input.parentNode.removeChild(
-        input.parentNode.querySelector(".error-text")
-      );
-    }
-  },
-};
-
-export const storeUtils = {
-  addStore: (storeProps) => {
-    const list = document.createElement("li");
-    list.classList.add("restaurant");
-    const store = Store(storeProps);
-    list.innerHTML = store;
-    querySelector(".restaurant-list").appendChild(list);
-  },
-
-  updateStore: (storeList, e) => {
-    const newStore = storeUtils.createStore(e);
-
-    try {
-      e.preventDefault();
-
-      validate.emptySelector(newStore.category);
-      validate.nameLength(newStore.name);
-      validate.emptySelector(newStore.dist);
-      validate.descLength(newStore.description);
-      validate.linkForm(newStore.link);
-
-      storeList.updateList(newStore);
-      storeUtils.addStore(newStore);
-
-      modalUtils.closeModal();
-    } catch (error) {
-      storeUtils.checkRequired("category", newStore.category, error);
-      storeUtils.checkRequired("name", newStore.name, error);
-      storeUtils.checkRequired("distance", newStore.dist, error);
-    }
-  },
-
-  checkRequired: (input, value, error) => {
-    if (value === "") {
-      const input = querySelector(`#${input}`);
-      modalUtils.addErrorText(input, error);
-    }
-  },
-
-  createStore: (e) => {
-    const data = new FormData(e.target);
-    return {
-      category: data.get("category"),
-      name: data.get("name"),
-      dist: data.get("distance"),
-      description: data.get("description"),
-      link: data.get("link"),
-    };
-  },
-
-  removeStoreElements: () => {
-    querySelector(".restaurant-list").replaceChildren();
-  },
-};
